requirements.txt
requests==2.28.1
alembic==1.10.1
sqlalchemy==2.0.0
# psycopg2==2.9.5
psycopg2-binary==2.9.5
pandas==2.0.1
numpy==1.24.3
python-dotenv==1.0.0
fastapi[standard]==0.112.2
pydantic>=2.7.0,<3.0.0
uvicorn
supabase

